requirements.txt
pandas==1.3.4
scv
scikit-allel==1.3.5
seaborn==0.11.2
numpy==1.21.3
matplotlib==3.5.1
argparse